refactor(tools): extract analyze_directory and analyze_file into tools/ (#1188)
